@@ -3,7 +3,6 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "develop": "nodemon --ignore dist --ext ts,js,json src/index.ts",
     "tsnode": "ts-node --esm src/index.mts",
     "build": "tsc --build --clean && tsc",
     "start": "node ./dist/src/index.mjs",
@@ -14,9 +13,6 @@
   "dependencies": {
     "@garmin/fitsdk": "^21.126.0",
     "@simplewebauthn/server": "^7.2.0",
-    "@tsconfig/node18": "^18.2.2",
-    "@types/connect-pg-simple": "^7.0.0",
-    "@types/pg": "^8.10.0",
     "connect-pg-simple": "^9.0.0",
     "cors": "^2.8.5",
     "debug": "^4.3.4",
@@ -26,7 +22,7 @@
     "express-validator": "^7.0.1",
     "formidable": "^3.5.1",
     "geographiclib-geodesic": "^2.0.0",
-    "helmet": "^7.0.0",
+    "helmet": "^7.1.0",
     "lodash": "^4.17.21",
     "luxon": "^3.4.4",
     "morgan": "^1.10.0",
@@ -34,15 +30,18 @@
   },
   "devDependencies": {
     "@simplewebauthn/typescript-types": "^7.0.0",
+    "@tsconfig/node18": "^18.2.2",
+    "@types/connect-pg-simple": "^7.0.0",
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
     "@types/express-session": "^1.17.7",
     "@types/formidable": "^3.4.2",
-    "@types/jest": "^29.5.1",
+    "@types/jest": "^29.5.10",
     "@types/lodash": "^4.14.202",
     "@types/luxon": "^3.3.4",
     "@types/morgan": "^1.9.4",
     "@types/node": "^20.4.4",
+    "@types/pg": "^8.10.0",
     "@types/supertest": "^2.0.12",
     "@typescript-eslint/eslint-plugin": "^6.1.0",
     "@typescript-eslint/parser": "^6.1.0",
@@ -54,12 +53,12 @@
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.1.1",
-    "jest": "^29.5.0",
+    "jest": "^29.7.0",
     "nodemon": "^3.0.1",
     "supertest": "^6.3.3",
-    "ts-jest": "^29.1.0",
+    "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
-    "typescript": "^5.0.4",
+    "typescript": "^5.3.2",
     "typescript-cp": "^0.1.9"
   }
 }
