@@ -1,12 +1,16 @@
+import pg from 'pg';
+import request from 'supertest';
+
 import app from '../../src/app.js';
+import getSchema from '../../src/lib/getSchema.js';
+import { isAuthenticated } from '../../src/routes/auth/auth.js';
 
-const request = require('supertest');
-const { Pool } = require('pg');
+jest.mock('../../src/routes/auth/auth');
+jest.mock('../../src/lib/getSchema.js')
 
-jest.mock('pg');
 
-jest.mock('../../src/lib/getSchema');
-const mockGetSchema = require('../../src/lib/getSchema');
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const dummy = isAuthenticated;
 
 const mockTrack1 = {
   id: '2',
@@ -17,16 +21,35 @@ const mockTrack1 = {
   ascent: 3,
 };
 
-jest
-  .spyOn(Pool.prototype, 'query')
-  .mockResolvedValue({
-    rows: [mockTrack1],
-  });
+
+const mockGetSchema = jest.mocked(getSchema)
+
+jest.mock('pg', () => {
+  const mClient = {
+    connect: jest.fn(),
+    query: jest.fn().mockResolvedValue([77]),
+    end: jest.fn(),
+  };
+  return { Pool: jest.fn(() => mClient) };
+});
+const mockedPg = jest.mocked(pg)
+
+
 
 describe('tracks - getall', () => {
-  beforeEach(
-    mockGetSchema.mockReset(),
-  );
+
+  beforeEach(() => {
+    mockGetSchema.mockReset()
+
+    const pool = new pg.Pool()
+    const mockedPool = jest.mocked(pool)
+    mockedPool.query.mockImplementation((...args: any): any => {
+      return {
+        rows: [mockTrack1]
+      }
+    })
+
+  });
   test('correctsid', async () => {
     mockGetSchema.mockResolvedValue('myschema');
     const response = await request(app)
